Refine flowgraph nodes for readability
- delete compaction tasks from executor while dropping

See also: #19780
