+++ internal/datanode/flow_graph_insert_buffer_node_test.go
@@ -253,94 +253,6 @@ func TestFlowGraphInsertBufferNode_Operate(t *testing.T) {
 	iBNode.flushManager = fm
 }
 
-/*
-func TestFlushSegment(t *testing.T) {
-	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
-	defer cancel()
-	idAllocMock := NewAllocatorFactory(1)
-	mockMinIO := memkv.NewMemoryKV()
-	insertChannelName := "datanode-02-test-flushsegment"
-
-	segmentID, _ := idAllocMock.allocID()
-	partitionID, _ := idAllocMock.allocID()
-	collectionID, _ := idAllocMock.allocID()
-	fmt.Printf("generate segmentID, partitionID, collectionID: %v, %v, %v\n",
-		segmentID, partitionID, collectionID)
-
-	collMeta := genCollectionMeta(collectionID, "test_flush_segment_txn")
-	flushMap := sync.Map{}
-	mockRootCoord := &RootCoordFactory{}
-
-	channel := newChannel("channel", collMeta.ID, collMeta.Schema, mockRootCoord)
-
-	err = channel.addNewSegment(segmentID, collMeta.ID, 0, insertChannelName, &internalpb.MsgPosition{}, &internalpb.MsgPosition{})
-	require.NoError(t, err)
-	channel.updateSegmentEndPosition(segmentID, &internalpb.MsgPosition{ChannelName: "TestChannel"})
-
-	finishCh := make(chan segmentFlushUnit, 1)
-
-	insertData := &InsertData{
-		Data: make(map[storage.FieldID]storage.FieldData),
-	}
-	insertData.Data[0] = &storage.Int64FieldData{
-		NumRows: []int64{10},
-		Data:    []int64{0, 1, 2, 3, 4, 5, 6, 7, 8, 9},
-	}
-	insertData.Data[1] = &storage.Int64FieldData{
-		NumRows: []int64{10},
-		Data:    []int64{1, 2, 3, 4, 5, 6, 7, 8, 9, 10},
-	}
-	insertData.Data[107] = &storage.FloatFieldData{
-		NumRows: []int64{10},
-		Data:    make([]float32, 10),
-	}
-	flushMap.Store(segmentID, insertData)
-
-	factory := msgstream.NewPmsFactory()
-	m := map[string]interface{}{
-		"receiveBufSize": 1024,
-		"pulsarAddress":  Params.PulsarAddress,
-		"pulsarBufSize":  1024}
-	err = factory.SetParams(m)
-	assert.Nil(t, err)
-	flushChan := make(chan flushMsg, 100)
-
-	memkv := memkv.NewMemoryKV()
-
-	fm := NewRendezvousFlushManager(&allocator{}, memkv, channel, func(*segmentFlushPack) error {
-		return nil
-	})
-
-	c := &nodeConfig{
-		channel:      channel,
-		factory:    factory,
-		allocator:    NewAllocatorFactory(),
-		vChannelName: "string",
-	}
-	ibNode, err := newInsertBufferNode(ctx, flushChan, fm, newCache(), c)
-	require.NoError(t, err)
-
-	flushSegment(collMeta,
-		segmentID,
-		partitionID,
-		collectionID,
-		&flushMap,
-		mockMinIO,
-		finishCh,
-		nil,
-		ibNode,
-		idAllocMock)
-
-	fu := <-finishCh
-	assert.NotNil(t, fu.field2Path)
-	assert.Equal(t, fu.segID, segmentID)
-
-    k := JoinIDPath(collectionID, partitionID, segmentID, 0)
-	key := path.Join(Params.StatsBinlogRootPath, k)
-	_, values, _ := mockMinIO.LoadWithPrefix(key)
-	assert.Equal(t, len(values), 1)
-}*/
-
 func genCollectionMeta(collectionID UniqueID, collectionName string) *etcdpb.CollectionMeta {
 	sch := schemapb.CollectionSchema{
 		Name:        collectionName,
@@ -976,95 +888,6 @@ func TestInsertBufferNode_updateSegmentStates(te *testing.T) {
 
 		assert.Error(te, err)
 		assert.Empty(te, seg)
-	}
-}
-
-func TestInsertBufferNode_BufferData(te *testing.T) {
-	Params.DataNodeCfg.FlushInsertBufferSize = 16 * (1 << 20) // 16 MB
-
-	tests := []struct {
-		isValid bool
-
-		indim         int64
-		expectedLimit int64
-
-		description string
-	}{
-		{true, 1, 4194304, "Smallest of the DIM"},
-		{true, 128, 32768, "Normal DIM"},
-		{true, 32768, 128, "Largest DIM"},
-		{false, 0, 0, "Illegal DIM"},
-	}
-
-	for _, test := range tests {
-		te.Run(test.description, func(t *testing.T) {
-			idata, err := newBufferData(test.indim)
-
-			if test.isValid {
-				assert.NoError(t, err)
-				assert.NotNil(t, idata)
-
-				assert.Equal(t, test.expectedLimit, idata.limit)
-				assert.Zero(t, idata.size)
-
-				capacity := idata.effectiveCap()
-				assert.Equal(t, test.expectedLimit, capacity)
-			} else {
-				assert.Error(t, err)
-				assert.Nil(t, idata)
-			}
-		})
-	}
-}
-
-func TestInsertBufferNode_BufferData_updateTimeRange(t *testing.T) {
-	Params.DataNodeCfg.FlushInsertBufferSize = 16 * (1 << 20) // 16 MB
-
-	type testCase struct {
-		tag string
-
-		trs        []TimeRange
-		expectFrom Timestamp
-		expectTo   Timestamp
-	}
-
-	cases := []testCase{
-		{
-			tag:        "no input range",
-			expectTo:   0,
-			expectFrom: math.MaxUint64,
-		},
-		{
-			tag: "single range",
-			trs: []TimeRange{
-				{timestampMin: 100, timestampMax: 200},
-			},
-			expectFrom: 100,
-			expectTo:   200,
-		},
-		{
-			tag: "multiple range",
-			trs: []TimeRange{
-				{timestampMin: 150, timestampMax: 250},
-				{timestampMin: 100, timestampMax: 200},
-				{timestampMin: 50, timestampMax: 180},
-			},
-			expectFrom: 50,
-			expectTo:   250,
-		},
-	}
-
-	for _, tc := range cases {
-		t.Run(tc.tag, func(t *testing.T) {
-			bd, err := newBufferData(16)
-			require.NoError(t, err)
-			for _, tr := range tc.trs {
-				bd.updateTimeRange(tr)
-			}
-
-			assert.Equal(t, tc.expectFrom, bd.tsFrom)
-			assert.Equal(t, tc.expectTo, bd.tsTo)
-		})
 	}
 }
